Make rinari-test work when test file requires 'test_helper' without a relative path.
Thanks to github user ptzn for the problem report and partial patch.

The latest Rails generators output "require 'test_helper'" at the top of the file, so we
need to add RAILS_ROOT/test to the library path when calling ruby-compilation-run. The neatest
way to do this was adding an optional 'ruby-options' parameter to that function, then passing
along a "-I...".